LibWeb: Move window.open TokenizedFeature parsing into its own file
