+++ Meta/gn/secondary/Userland/Libraries/LibWeb/HTML/BUILD.gn
@@ -163,6 +163,7 @@ source_set("HTML") {
     "TimeRanges.cpp",
     "Timer.cpp",
     "ToggleEvent.cpp",
+    "TokenizedFeatures.cpp",
     "TrackEvent.cpp",
     "TraversableNavigable.cpp",
     "UserActivation.cpp",
